Update logs table
