+++ ui/src/components/BasicBoiler/index.tsx
@@ -2,17 +2,23 @@ import { ReactNode } from "react";
 
 type Props = {
   className?: string;
+  textClassName?: string;
   text: string;
   icon?: ReactNode;
 };
 
-const BasicBoiler: React.FC<Props> = ({ className, text, icon }) => {
+const BasicBoiler: React.FC<Props> = ({
+  className,
+  textClassName,
+  text,
+  icon,
+}) => {
   return (
     <div
-      className={`flex w-full flex-1 items-center justify-center ${className}`}>
+      className={`flex w-full flex-1 items-center justify-center text-xl ${className}`}>
       <div className="flex flex-col items-center gap-6">
         {icon}
-        <p className="text-xl font-thin">{text}</p>
+        <p className={`font-thin ${textClassName}`}>{text}</p>
       </div>
     </div>
   );
